Move audit functions to new import method
